# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Desafio Sicredi - Documentação",
        "description": "Documentação da API de Autenticação e Associados",
        "contact": {
            "email": "guilherme@api.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registrar novo usuário",
                "description": "Cria um novo usuário no sistema",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "description": "Dados do usuário para registro",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Fazer login",
                "description": "Autentica um usuário e retorna um token JWT",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "description": "Credenciais do usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Fazer logout",
                "description": "Invalida o token JWT do usuário autenticado",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Obter usuário autenticado",
                "description": "Retorna os dados do usuário atualmente autenticado",
                "operationId": "b3c2811c57c1ffa70c76a938f23f36a6",
                "responses": {
                    "200": {
                        "description": "Dados do usuário retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "example": "joao.silva@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "Nome completo do usuário",
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "description": "Email do usuário",
                        "type": "string",
                        "format": "email",
                        "example": "joao.silva@example.com"
                    },
                    "password": {
                        "description": "Senha do usuário (mínimo 6 caracteres)",
                        "type": "string",
                        "format": "password",
                        "example": "senha123"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Email do usuário",
                        "type": "string",
                        "format": "email",
                        "example": "joao.silva@example.com"
                    },
                    "password": {
                        "description": "Senha do usuário",
                        "type": "string",
                        "format": "password",
                        "example": "senha123"
                    }
                },
                "type": "object"
            },
...
            "RegisterResponse": {
                "properties": {
                    "token": {
                        "description": "Token JWT de autenticação",
                        "type": "string",
                        "example": "eyJ0132sOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "properties": {
                    "token": {
                        "description": "Token JWT de autenticação",
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    },
                    "expires_in": {
                        "description": "Tempo de expiração do token em segundos",
                        "type": "integer",
                        "example": 3600
                    }
                },
                "type": "object"
            },
            "MessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Successfully logged out"
                    }
                },
                "type": "object"
            },
...
            "ApiError": {
                "properties": {
                    "path": {
                        "type": "string",
                        "example": "/api/auth/login"
                    },
                    "method": {
                        "type": "string",
                        "example": "POST"
                    },
                    "status": {
                        "type": "integer",
                        "example": 401
                    },
                    "message": {
                        "type": "string",
                        "example": "Invalid credentials!"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "path": {
                        "type": "string",
                        "example": "/api/auth/register"
                    },
                    "method": {
                        "type": "string",
                        "example": "POST"
                    },
                    "status": {
                        "type": "integer",
                        "example": 422
                    },
                    "message": {
                        "type": "string",
                        "example": "Invalid input data."
                    },
                    "errors": {
                        "description": "Erros de validação por campo",
                        "type": "object",
                        "example": {
                            "email": [
                                "The email has already been taken."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT no formato: Bearer {seu-token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}